// apps/web/src/features/mechanical-integrity/types/impairment.types.ts
export type MiImpairmentSummary = {
  totalRecords: number;
  activeBypasses: number;
  activeImpairments: number;
  activeOverrides: number;
  activeInhibits: number;
  expiredBypasses: number;
  pendingApproval: number;
  pendingRestorationVerification: number;
  criticalSafeguardsImpaired: number;
  lopaSilIplsImpaired: number;
  psvReliefDevicesImpaired: number;
  sifsImpaired: number;
  interlocksImpaired: number;
  criticalAlarmsImpaired: number;
  startupBlocked: number;
  mocRequiredSuggested: number;
  ptwLotoLinked: number;
  restoredThisMonth: number;
};